html/semantics/embedded-content-0/media-elements/loading-the-media-resource/007.html

<!doctype html><base href=http://w3c-test.org>
<title>NOT invoking resource selection by inserting &lt;source> when networkState is not NETWORK_EMPTY</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>var loadstartCount = 0</script>
<video onloadstart=loadstartCount++><source></video><!-- starts resource selection and sets networkState to NETWORK_NO_SOURCE -->
<script>
async_test(function(t) {
  var v = document.querySelector('video');
  v.appendChild(document.createElement('source'));
  window.onload = t.step_func(function() {
    assert_equals(loadstartCount, 1);
    t.done();
  });
});
</script>
